requirements/base.txt
Django==5.1.1
djangorestframework==3.15.2
moviepy==1.0.3
pillow==10.4.0
djangorestframework-simplejwt==5.3.1